Split backend observability event contracts
